% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R
\docType{class}
\name{ADsparse}
\alias{ADsparse}
\alias{adsparse}
\title{AD sparse matrix class}
\description{
Sparse matrices in \pkg{RTMB} are essentially \code{dgCMatrix} with an \code{advector} x-slot.
}
\section{Slots}{

\describe{
\item{\code{x}}{Non-zeros}

\item{\code{i}}{row indices (zero based)}

\item{\code{p}}{col pointers (zero based)}

\item{\code{Dim}}{Dimension}
}}

